configured buildPrebidMobile.sh for circle ci build

+++ scripts/buildPrebidMobile.sh
@@ -11,7 +11,7 @@
 
 # Merge Script
 if [ -d "scripts" ]; then
-cd scripts/
+  cd scripts/
 fi
 
 set -e
@@ -24,11 +24,15 @@ RED='\033[0;31m'
 GREEN='\033[0;32m'
 NC='\033[0m' # No Color
 
-function echoX {
-echo -e "PREBID BUILDLOG: $@"
+function echoX() {
+  echo -e "PREBID BUILDLOG: $@"
 }
 
-die() { echoX "$@" 1>&2 ; echoX "End Script"; exit 1;  }
+die() {
+  echoX "$@" 1>&2
+  echoX "End Script"
+  exit 1
+}
 
 ######################
 # Build Settings
@@ -50,13 +54,11 @@ echoX "$BASEDIR"
 # set the default release version to what's in the project's build.gradle file
 RELEASE_VERSION=""
 regex="prebidVersionName.*=.*\"(.*)\""
-while read -r line;
-do
-if [[ $line =~ $regex ]];
-then
-RELEASE_VERSION=${BASH_REMATCH[1]};
-fi
-done < $LIBDIR/build.gradle
+while read -r line; do
+  if [[ $line =~ $regex ]]; then
+    RELEASE_VERSION=${BASH_REMATCH[1]}
+  fi
+done <$LIBDIR/build.gradle
 
 echoX "Start building Prebid Mobile version $RELEASE_VERSION"
 
@@ -90,71 +92,71 @@ projectPaths=(
   "$BASEDIR/PrebidMobile/PrebidMobile-admobAdapters"
   "$BASEDIR/PrebidMobile/PrebidMobile-maxAdapters"
 )
-
 mkdir "$OUTDIR/aar"
 for n in ${!modules[@]}; do
 
-	echo -e "\n"
-	echoX "Assembling ${modules[$n]}"
-	cd $LIBDIR
-	# clean existing build results, exclude test task, and assemble new release build
-	(./gradlew -i --no-daemon ${modules[$n]}:assembleRelease > $LOGPATH/build.log 2>&1 || die "Build failed, check log in $LOGPATH/build.log" )
+  echo -e "\n"
+  echoX "Assembling ${modules[$n]}"
+  cd $LIBDIR
+  # clean existing build results, exclude test task, and assemble new release build
+  (./gradlew -i --no-daemon ${modules[$n]}:assembleRelease >$LOGPATH/build.log 2>&1 || die "Build failed, check log in $LOGPATH/build.log")
 
-    # Make folder generated/temp/output
-	echoX "Packaging ${modules[$n]}"
-	mkdir $TEMPDIR
-	cd $TEMPDIR
-	mkdir output
-	
-	AARPATH_ABSOLUTE="${projectPaths[$n]}/$AARPATH"
+  # Make folder generated/temp/output
 
-	cd $AARPATH_ABSOLUTE
-	cp ${modules[$n]}-release.aar $OUTDIR/aar
-	unzip -q -o ${modules[$n]}-release.aar
-	cd $TEMPDIR/output
+    echoX "Packaging ${modules[$n]}"
+    mkdir $TEMPDIR
+    cd $TEMPDIR
+    mkdir output
 
-	# Extracting the Contents of a JAR File
-	jar xf $AARPATH_ABSOLUTE/classes.jar
-	rm $AARPATH_ABSOLUTE/classes.jar
+    AARPATH_ABSOLUTE="${projectPaths[$n]}/$AARPATH"
 
-	# Handle ProGuard rules from .aar into .jar
-	# rename proguard.txt to proguard.pro
-	mv $AARPATH_ABSOLUTE/proguard.{txt,pro}
-	mkdir -p $AARPATH_ABSOLUTE/META-INF
-	mkdir $AARPATH_ABSOLUTE/META-INF/proguard
-	mv $AARPATH_ABSOLUTE/proguard.pro $AARPATH_ABSOLUTE/META-INF/proguard
-	# move META-INF into a result direcotory
-	mv $AARPATH_ABSOLUTE/META-INF $TEMPDIR/output
+    cd $AARPATH_ABSOLUTE
+    cp ${modules[$n]}-release.aar $OUTDIR/aar
+    unzip -q -o ${modules[$n]}-release.aar
+    cd $TEMPDIR/output
 
-	rm -r $TEMPDIR/output/META-INF/com
+    # Extracting the Contents of a JAR File
+    jar xf $AARPATH_ABSOLUTE/classes.jar
+    rm $AARPATH_ABSOLUTE/classes.jar
 
-	# Creating a JAR File
+    # Handle ProGuard rules from .aar into .jar
+    # rename proguard.txt to proguard.pro
+    mv $AARPATH_ABSOLUTE/proguard.{txt,pro}
+    mkdir -p $AARPATH_ABSOLUTE/META-INF
+    mkdir $AARPATH_ABSOLUTE/META-INF/proguard
+    mv $AARPATH_ABSOLUTE/proguard.pro $AARPATH_ABSOLUTE/META-INF/proguard
+    # move META-INF into a result direcotory
+    mv $AARPATH_ABSOLUTE/META-INF $TEMPDIR/output
 
-  if [ "${modules[$n]}" == "PrebidMobile-maxAdapters" ]; then
-    jar cf ${modules[$n]}.jar org* com* META-INF*
-  else
-    jar cf ${modules[$n]}.jar org* META-INF*
+    rm -r $TEMPDIR/output/META-INF/com
+
+    # Creating a JAR File
+  if [ "$1" != "-nojar" ]; then
+    if [ "${modules[$n]}" == "PrebidMobile-maxAdapters" ]; then
+      jar cf ${modules[$n]}.jar org* com* META-INF*
+    else
+      jar cf ${modules[$n]}.jar org* META-INF*
+    fi
+
+    # move jar into a result direcotory
+    mv ${modules[$n]}.jar $OUTDIR
+
+    cd $LIBDIR
+
+    # Javadoc
+    echoX "Preparing ${modules[$n]} Javadoc"
+    ./gradlew -i --no-daemon ${modules[$n]}:javadocJar >$LOGPATH/javadoc.log 2>&1 || die "Build Javadoc failed, check log in $LOGPATH/javadoc.log"
+
+    # Sources
+    echoX "Preparing ${modules[$n]} Sources"
+    ./gradlew -i --no-daemon ${modules[$n]}:sourcesJar >$LOGPATH/sources.log 2>&1 || die "Build Sources failed, check log in $LOGPATH/sources.log"
+
+    # copy sources and javadoc into a result direcotory
+    BUILD_LIBS_PATH_ABSOLUTE="${projectPaths[$n]}/$BUILD_LIBS_PATH"
+    cp -a $BUILD_LIBS_PATH_ABSOLUTE/. $OUTDIR/
+    # clean tmp dir
   fi
-
-	# move jar into a result direcotory
-	mv ${modules[$n]}.jar $OUTDIR
-
-	cd $LIBDIR
-
-	# Javadoc
-	echoX "Preparing ${modules[$n]} Javadoc"
-	./gradlew -i --no-daemon ${modules[$n]}:javadocJar>$LOGPATH/javadoc.log 2>&1 || die "Build Javadoc failed, check log in $LOGPATH/javadoc.log"
-
-	# Sources
-	echoX "Preparing ${modules[$n]} Sources"
-	./gradlew -i --no-daemon ${modules[$n]}:sourcesJar>$LOGPATH/sources.log 2>&1 || die "Build Sources failed, check log in $LOGPATH/sources.log"
-
-	# copy sources and javadoc into a result direcotory
-	BUILD_LIBS_PATH_ABSOLUTE="${projectPaths[$n]}/$BUILD_LIBS_PATH"
-	cp -a $BUILD_LIBS_PATH_ABSOLUTE/. $OUTDIR/
-
-	# clean tmp dir
-	rm -r $TEMPDIR
+  rm -r $TEMPDIR
 done
 
 ### omsdk
@@ -183,24 +185,25 @@ echoX "Preparing fat PrebidDemo library"
 cd $OUTDIR
 mkdir $TEMPDIR
 
-cd $TEMPDIR; 
+cd $TEMPDIR
 
 unzip -qq -uo $OUTDIR/omsdk.jar
-unzip -qq -uo $OUTDIR/PrebidMobile.jar
-unzip -qq -uo $OUTDIR/PrebidMobile-core.jar
+if [ "$1" != "-nojar" ]; then
+  unzip -qq -uo $OUTDIR/PrebidMobile.jar
+  unzip -qq -uo $OUTDIR/PrebidMobile-core.jar
 
-# unzip second proguard
-unzip -qq -B $OUTDIR/PrebidMobile.jar "META-INF/proguard/proguard.pro"
+  # unzip second proguard
+  unzip -qq -B $OUTDIR/PrebidMobile.jar "META-INF/proguard/proguard.pro"
 
-# append text from second proguard
-cat "$TEMPDIR/META-INF/proguard/proguard.pro~" >> "$TEMPDIR/META-INF/proguard/proguard.pro"
-rm "$TEMPDIR/META-INF/proguard/proguard.pro~"
+  cat "$TEMPDIR/META-INF/proguard/proguard.pro~" >>"$TEMPDIR/META-INF/proguard/proguard.pro"
+  rm "$TEMPDIR/META-INF/proguard/proguard.pro~"
 
-rm $TEMPDIR/org/prebid/mobile/core/BuildConfig.class
-jar -cvf PrebidMobile.jar -C $TEMPDIR .
+  rm $TEMPDIR/org/prebid/mobile/core/BuildConfig.class
+  jar -cvf PrebidMobile.jar -C $TEMPDIR .
 
-mkdir $FAT_PATH
-mv PrebidMobile.jar $FAT_PATH
+  mkdir $FAT_PATH
+  mv PrebidMobile.jar $FAT_PATH
+fi
 
 rm -r $TEMPDIR
